mod_customcert Fix bug due to totara fastpath in get_in_or_equal
Totara's get in or equal implementation has a "fast path" for integer
parameters, instead of using param substitution and returning the params
array it explodes the integers into an explicit IN query and returns an
empty array.

The mod customcert was assuming that a lack of a conditions param from
this get in or equal call indicated a empty result and shortcutting the
behaviour. This lead to the behaviour of shortcutting incorrectly when
installed on totara.

To fix the behaviour, we simply avoid the the shortcutting and make the
SQL parameter call anyway, resolving the underlying issue.

WR386270